Rename CivicrmGroups namespace to Groups
