+++ decidim-civicrm/app/views/decidim/civicrm/verifications/groups/admin/groups/index.html.erb
