+++ decidim-civicrm/lib/decidim/civicrm.rb
@@ -1,9 +1,9 @@
 # frozen_string_literal: true
 
-require "decidim/civicrm/verifications/civicrm_groups/engine"
-require "decidim/civicrm/verifications/civicrm_groups/admin"
-require "decidim/civicrm/verifications/civicrm_groups/admin_engine"
-require "decidim/civicrm/verifications/civicrm_groups/workflow"
+require "decidim/civicrm/verifications/groups/engine"
+require "decidim/civicrm/verifications/groups/admin"
+require "decidim/civicrm/verifications/groups/admin_engine"
+require "decidim/civicrm/verifications/groups/workflow"
 
 module Decidim
   # This namespace holds the logic of the `Civicrm` component. This component
