+++ decidim-civicrm/lib/decidim/civicrm/verifications/groups/engine.rb
@@ -3,10 +3,10 @@
 module Decidim
   module Civicrm
     module Verifications
-      module CivicrmGroups
+      module Groups
         # This is an engine that performs an example user authorization.
         class Engine < ::Rails::Engine
-          isolate_namespace Decidim::Civicrm::Verifications::CivicrmGroups
+          isolate_namespace Decidim::Civicrm::Verifications::Groups
 
           paths["db/migrate"] = nil
           paths["lib/tasks"] = nil
